requirements.txt
faiss-cpu==1.12.0
Flask==3.1.2
langchain==0.3.27
langchain-community==0.3.30
llama-cpp-python==0.2.90
sentence-transformers==2.7.0
textual==0.70.0
urwid==3.0.3